backend/requirements.txt
anyio==4.2.0
attrs==23.2.0
Babel==2.10.3
bcrypt==3.2.2
blinker==1.7.0
certifi==2023.11.17
chardet==5.2.0
click==8.1.6
cryptography==41.0.7
distro==1.9.0
docker==5.0.3
# docker-compose removed due to PyYAML conflict
docopt==0.6.2
fasteners==0.18
h11==0.14.0
httpcore==1.0.2
httplib2==0.20.4
httpx==0.26.0
idna==3.6
importlib-metadata==4.12.0
Jinja2==3.1.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
Mako==1.3.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
monotonic==1.6
more-itertools==10.2.0
oauthlib==3.2.2
packaging==24.0
paramiko==2.12.0
pexpect==4.9.0
pillow==10.2.0
ptyprocess==0.7.0
Pygments==2.17.2
PyJWT==2.7.0
PyNaCl==1.5.0
pyparsing==3.1.1
pyrsistent==0.20.0
pyserial==3.5
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
setuptools==68.1.2
six==1.16.0
sniffio==1.3.0
socksio==1.0.0
texttable==1.6.7
tomli==2.0.1
typer==0.9.0
typing_extensions==4.10.0
urllib3==2.0.7
websocket-client==1.7.0
wheel==0.42.0
zipp==1.0.0
Flask
flask-cors
flask-session
werkzeug
supabase

marshmallow
# requests is already present